Split setting definitions from setting accessors [refactor] (#334)
This will permit background.js to load the setting definitions even when
it can no longer use the accessors.
