+++ app/bg/background.js
@@ -14,7 +14,7 @@ if(false) { // Vendor files - listed here only so they'll be bundled
     require('process/browser');
 }
 
-const S = require('common/setting-definitions');    // in app/
+const S = require('common/setting-accessors');    // in app/
 
 /// The module exports, for use in command-line debugging
 let me = {
